Enhancement Added Location and Path fields to both File and Folder Fix Regenerate a new form after submit save Enhancement Added onReplaceUrl callback/hook in AssetAdmin to update the url without an entry made in history
